adding auth and cors to cloud functions

+++ delete-entry.py
@@ -1,7 +1,21 @@
 import sqlalchemy
 import os
-from flask import jsonify
+import random
+import functions_framework
+import json
 from datetime import datetime
+from google.oauth2 import id_token
+from google.auth.transport import requests
+
+
+def verify_token(request) -> bool:
+    token = request.args.get('token')
+    try:
+        idinfo = id_token.verify_oauth2_token(token, requests.Request())
+        return True
+
+    except ValueError:
+        return False
 
 def get_db():
     connection_name = os.environ.get('CONNECTION_NAME')
@@ -28,14 +42,35 @@ def get_db():
     return db
 
 def main(request):
+    # Set CORS headers for the preflight request
+    if request.method == 'OPTIONS':
+        # Allows GET requests from any origin with the Content-Type
+        # header and caches preflight response for an 3600s
+        headers = {
+            'Access-Control-Allow-Origin': '*',
+            'Access-Control-Allow-Methods': 'GET, POST, PUT, DELETE',
+            'Access-Control-Allow-Headers': 'Content-Type',
+            'Access-Control-Max-Age': '3600'
+        }
+
+        return ('', 204, headers)
+
+    # Set CORS headers for the main request
+    headers = {
+        'Access-Control-Allow-Origin': '*',
+        'Content-Type': 'application/json'
+    }
+
+    if not verify_token(request):
+        return ('unauthenticated', 401, headers)
+
     db = get_db()
 
     entryId = request.args.get('entryId')
-    entryId = int(entryId) if entryId else entryId
 
     if entryId:
         query_string = """
-            DELETE FROM entries WHERE "entryId" = :entryId
+            DELETE FROM entries WHERE "entryId" = :entryId;
         """
 
         query = sqlalchemy.text(query_string)
@@ -43,9 +78,10 @@ def main(request):
         try:
             with db.connect() as conn:
                 db.execute(query, entryId = entryId)  
-                return jsonify(success = True)
+
+                return (f"successfully deleted entry {entryId}.", 200, headers)
             
         except Exception as e:
-            return 'Error: {}'.format(str(e))
+            return ('Error: {}'.format(str(e)), 400, headers)
     
-    return "Record not found", 400
+    return ("Record not found".format(str(e)), 400, headers)
